Revert "MAIS A JOUR"
This reverts commit 7b944a2715ed43313f8c0cef1672139fde98a437.

+++ gtsco2/forms/GSTnote/gtevalustion.cs
@@ -414,20 +414,20 @@ namespace gtsco2.forms
 
                             if (row["Control1"].ToString() != "" )
                             {
-                                ev.Control_1 = double.Parse(row["Control1"].ToString());
+                                ev.Control_1 = float.Parse(row["Control1"].ToString());
                             }
                             if (row["Control2"].ToString() != "")
                             {
-                                ev.Control_2 = double.Parse(row["Control2"].ToString());
+                                ev.Control_2 = float.Parse(row["Control2"].ToString());
                             }
                             if (row["Exam"].ToString() != "")
                             {
-                                ev.Exam = double.Parse(row["Exam"].ToString());
+                                ev.Exam = float.Parse(row["Exam"].ToString());
                             }
                             
                                 if (row["Rattrapage"].ToString() != "")
                                 {
-                                    ev.Rattrapage = double.Parse(row["Rattrapage"].ToString());
+                                    ev.Rattrapage = float.Parse(row["Rattrapage"].ToString());
                                 }
 
                             //calcule de la moyenne de module
@@ -442,7 +442,7 @@ namespace gtsco2.forms
                                     ev.Moyenne_Module_ApRt = (double)((ev.Control_2 + ev.Control_1 + Math.Max((double)ev.Exam, ((double)ev.Rattrapage * 2))) / 4);
                                 }
                                 else { ev.Moyenne_Module_ApRt =  Math.Max((double)ev.Moyenne_Module_AvRt, (double)ev.Rattrapage ); }
-                                
+
 
                             }
                             else if (ev.Control_1 != null && ev.Control_2 == null && ev.Exam != null && ev.Rattrapage != null)
